Make class as tool, move the location
